WebApp/requirements.txt
Flask==3.0.0
flask_wtf==1.2.1
Requests==2.31.0
WTForms==3.1.1
gunicorn==20.1.0
Werkzeug==3.0.1
protobuf==4.25.1
Requests==2.31.0
google-api-core==2.12.0
google-auth==2.23.4
google-cloud-bigtable==2.17.0
google-cloud-core==2.3.2
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6